// src/model/character/pet-equipment.ts
// 장착 펫 정보
export interface CharacterPetEquipment {
  pet_1_name: string; // 펫 1 명
  pet_1_pet_type: string; // 펫 1 원더 펫 종류
  pet_1_date_expire: string; // 펫 1 마법의 시간 (UTC0)
  pet_2_name: string; // 펫 1 명
  pet_2_pet_type: string; // 펫 1 원더 펫 종류
  pet_2_date_expire: string; // 펫 1 마법의 시간 (UTC0)
  pet_3_name: string; // 펫 1 명
  pet_3_pet_type: string; // 펫 1 원더 펫 종류
  pet_3_date_expire: string; // 펫 1 마법의 시간 (UTC0)
}